requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.27.0
attrs==24.2.0
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.95.0
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
paramiko==3.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==1.10.2
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.26.1
typing_extensions==4.12.2
uvicorn==0.21.1
yarl==1.11.1
